Use llvm-openmp on Linux as well
When building with `mkl`, upstream build system uses the same OpenMP
implementation as `mkl`, which is `llvm-openmp`.  In generic variant,
the upstream can use either `llvm-openmp` or `libgomp`, but
`llvm-openmp` takes precedence if it is installed.  In CUDA-enabled
builds, `llvm-openmp` was installed as an indirect dependency,
and therefore the package used it rather than `libgomp`.
For consistency and to avoid need to patch the build system, just depend
on `llvm-openmp` for all Linux targets.