Fix inductor include and library paths for CUDA target
It looks like we now need to explicitly pass CUDA target include
and `libcuda.so` stub directories while building inductor.  Introduce
a substitution `@CUDA_TARGET@` value in the patch, and replace it with
appropriate target in `build.sh`.

This leaves the unsubstituted `@CUDA_TARGET@` path on Windows, but that
shouldn't do any harm -- the path will simply not exist.

Signed-off-by: Michał Górny <mgorny@quansight.com>